@@ -20,11 +20,13 @@
     "supabase:types": "supabase gen types typescript --project-id yrfszduvxwygylvlnoza > database/types.ts"
   },
   "dependencies": {
+    "@formkit/vue": "^1.6.9",
     "@supabase/supabase-js": "^2.43.4",
     "@tanstack/vue-table": "^8.21.2",
     "@vueuse/core": "^12.8.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "formkit": "^1.6.9",
     "iconify-icon": "^2.3.0",
     "lucide-vue-next": "^0.477.0",
     "pinia": "^3.0.1",
@@ -34,6 +36,7 @@
     "tailwindcss-animate": "^1.0.7",
     "unplugin-vue-components": "^28.4.1",
     "vue": "^3.5.13",
+    "vue-meta": "^3.0.0-alpha.7",
     "vue-router": "^4.5.0"
   },
   "devDependencies": {
@@ -42,14 +45,14 @@
     "@types/node": "^22.13.4",
     "@vitejs/plugin-vue": "^5.2.1",
     "@vue/eslint-config-prettier": "^10.2.0",
-    "@vue/eslint-config-typescript": "^14.4.0",
+    "@vue/eslint-config-typescript": "^14.5.0",
     "@vue/tsconfig": "^0.7.0",
     "autoprefixer": "^10.4.20",
-    "eslint": "^9.21.0",
-    "eslint-plugin-vue": "^9.32.0",
+    "eslint": "^9.22.0",
+    "eslint-plugin-vue": "^10.0.0",
     "jiti": "^2.4.2",
     "npm-run-all2": "^7.0.2",
-    "prettier": "^3.5.2",
+    "prettier": "^3.5.3",
     "supabase": "^2.15.8",
     "tailwindcss": "^3.4.17",
     "typescript": "~5.8.2",
