@@ -43,7 +43,7 @@ export type Database = {
         Row: {
           collaborators: string[]
           created_at: string
-          description: string | null
+          description: string
           id: number
           name: string
           slug: string
@@ -52,7 +52,7 @@ export type Database = {
         Insert: {
           collaborators?: string[]
           created_at?: string
-          description?: string | null
+          description?: string
           id?: never
           name: string
           slug: string
@@ -61,7 +61,7 @@ export type Database = {
         Update: {
           collaborators?: string[]
           created_at?: string
-          description?: string | null
+          description?: string
           id?: never
           name?: string
           slug?: string
@@ -77,6 +77,7 @@ export type Database = {
           due_date: string | null
           id: number
           name: string
+          profile_id: string
           project_id: number | null
           status: Database["public"]["Enums"]["current_status"]
         }
@@ -87,6 +88,7 @@ export type Database = {
           due_date?: string | null
           id?: never
           name: string
+          profile_id: string
           project_id?: number | null
           status?: Database["public"]["Enums"]["current_status"]
         }
@@ -97,10 +99,18 @@ export type Database = {
           due_date?: string | null
           id?: never
           name?: string
+          profile_id?: string
           project_id?: number | null
           status?: Database["public"]["Enums"]["current_status"]
         }
         Relationships: [
+          {
+            foreignKeyName: "tasks_profile_id_fkey"
+            columns: ["profile_id"]
+            isOneToOne: false
+            referencedRelation: "profiles"
+            referencedColumns: ["id"]
+          },
           {
             foreignKeyName: "tasks_project_id_fkey"
             columns: ["project_id"]
