@@ -1,40 +1,58 @@
 <script setup lang="ts">
+import { menuKey, type MenuInjectionOptions } from '@/utils/injectionKeys'
+
 interface LinkProp {
   title: string
   to?: string
   icon: string
 }
 
-defineProps<{ links: LinkProp[] }>()
+defineProps<{
+  links: LinkProp[]
+}>()
 
-const emits = defineEmits<{ actionClicked: [string] }>()
+const emits = defineEmits<{
+  actionClicked: [string]
+}>()
 
 const emitActionClicked = (linkTitle: string) => {
-  return emits('actionClicked', linkTitle)
+  emits('actionClicked', linkTitle)
 }
+
+const { menuOpen } = inject(menuKey) as MenuInjectionOptions
 </script>
 
 <template>
   <template v-for="link in links" :key="link.title">
     <RouterLink
       v-if="link.to"
-      exactActiveClass="tex-primary bg-muted"
-      :key="link.title"
+      exactActiveClass="text-primary bg-muted"
       :to="link.to"
       class="nav-link"
+      :class="{ 'justify-normal': menuOpen, 'justify-center': !menuOpen }"
     >
       <iconify-icon :icon="link.icon"></iconify-icon>
-      <span class="hidden lg:block text-nowrap">{{ link.title }}</span>
+      <span class="text-nowrap" :class="{ block: menuOpen, hidden: !menuOpen }">{{
+        link.title
+      }}</span>
     </RouterLink>
-    <div v-else class="nav-link" cursor-pointer @click="emitActionClicked(link.title)">
+
+    <div
+      v-else
+      class="nav-link cursor-pointer"
+      :class="{ 'justify-normal': menuOpen, 'justify-center': !menuOpen }"
+      @click="emitActionClicked(link.title)"
+    >
       <iconify-icon :icon="link.icon"></iconify-icon>
-      <span class="hidden lg:block text-nowrap">{{ link.title }}</span>
+      <span class="text-nowrap" :class="{ block: menuOpen, hidden: !menuOpen }">{{
+        link.title
+      }}</span>
     </div>
   </template>
 </template>
 
 <style scoped>
 .nav-link {
-  @apply flex items-center gap-3 px-4 py-2 mx-2 transition-colors rounded-lg hover:text-primary justify-center lg:justify-normal text-muted-foreground;
+  @apply flex items-center gap-3 px-4 py-2 mx-2 transition-colors rounded-lg hover:text-primary text-muted-foreground;
 }
 </style>
