@@ -1,91 +1,114 @@
 <script setup lang="ts">
-import { projectQuery } from '@/utils/supaQueries'
-import type { Project } from '@/utils/supaQueries'
+const { slug } = useRoute('/projects/[slug]').params
 
-const route = useRoute('/projects/[slug]')
-
-const project = ref<Project | null>(null)
+const projectsLoader = useProjectsStore()
+const { project } = storeToRefs(projectsLoader)
+const { getProject, updateProject } = projectsLoader
 
 watch(
   () => project.value?.name,
-  () => (usePageStore().pageData.title = `Project: ${project.value?.name || ''}`)
+  () => {
+    usePageStore().pageData.title = `Project: ${project.value?.name || ''}`
+  }
 )
 
-const getProjects = async () => {
-  const { data, error, status } = await projectQuery(route.params.slug)
+await getProject(slug)
 
-  if (error) if (error) useErrorStore().setError({ error, customCode: status })
+const { getProfilesByIds } = useCollabs()
 
-  project.value = data
-}
-
-await getProjects()
+const collabs = project.value?.collaborators
+  ? await getProfilesByIds(project.value?.collaborators)
+  : []
 </script>
 
 <template>
-  <Table v-if="project">
-    <TableRow>
-      <TableHead> Name </TableHead>
-      <TableCell> {{ project.name }} </TableCell>
-    </TableRow>
-    <TableRow>
-      <TableHead> Description </TableHead>
-      <TableCell>
-        {{ project.description }}
-      </TableCell>
-    </TableRow>
-    <TableRow>
-      <TableHead> Status </TableHead>
-      <TableCell>{{ project.status }}</TableCell>
-    </TableRow>
-    <TableRow>
-      <TableHead> Collaborators </TableHead>
-      <TableCell>
-        <div class="flex">
-          <Avatar
-            class="-mr-4 border border-primary hover:scale-110 transition-transform"
-            v-for="collab in project.collaborators"
-            :key="collab"
-          >
-            <RouterLink class="w-full h-full flex items-center justify-center" to="">
-              <AvatarImage src="" alt="" />
-              <AvatarFallback> </AvatarFallback>
-            </RouterLink>
-          </Avatar>
-        </div>
-      </TableCell>
-    </TableRow>
-  </Table>
+  <div>
+    <Table v-if="project">
+      <TableRow>
+        <TableHead> Name </TableHead>
+        <TableCell>
+          <AppInPlaceEditText v-model="project.name" @commit="updateProject" />
+        </TableCell>
+      </TableRow>
+      <TableRow>
+        <TableHead> Description </TableHead>
+        <TableCell>
+          <AppInPlaceEditTextarea
+            class="h-20"
+            v-model="project.description"
+            @commit="updateProject"
+          />
+        </TableCell>
+      </TableRow>
+      <TableRow>
+        <TableHead> Status </TableHead>
+        <TableCell>
+          <AppInPlaceEditStatus v-model="project.status" @commit="updateProject" />
+        </TableCell>
+      </TableRow>
+      <TableRow>
+        <TableHead> Collaborators </TableHead>
+        <TableCell>
+          <div class="flex">
+            <Avatar
+              class="-mr-4 border border-primary hover:scale-110 transition-transform"
+              v-for="collab in collabs"
+              :key="collab.id"
+            >
+              <RouterLink
+                class="w-full h-full flex items-center justify-center"
+                :to="{
+                  name: '/users/[username]',
+                  params: { username: collab.username }
+                }"
+              >
+                <AvatarImage :src="collab.avatar_url || ''" alt="" />
+                <AvatarFallback> </AvatarFallback>
+              </RouterLink>
+            </Avatar>
+          </div>
+        </TableCell>
+      </TableRow>
+    </Table>
 
-  <section v-if="project" class="mt-10 flex flex-col md:flex-row gap-5 justify-between grow">
-    <div class="flex-1">
-      <h2>Tasks</h2>
-      <div class="table-container">
-        <Table>
-          <TableHeader>
-            <TableRow>
-              <TableHead> Name </TableHead>
-              <TableHead> Status </TableHead>
-              <TableHead> Due Date </TableHead>
-            </TableRow>
-          </TableHeader>
-          <TableBody>
-            <TableRow v-for="task in project.tasks" :key="task.id">
-              <TableCell> Lorem ipsum dolor sit amet. </TableCell>
-              <TableCell> In progress </TableCell>
-              <TableCell> 22/08/2024 </TableCell>
-            </TableRow>
-          </TableBody>
-        </Table>
+    <section v-if="project" class="mt-10 flex flex-col md:flex-row gap-5 justify-between grow">
+      <div class="flex-1">
+        <h2>Tasks</h2>
+        <div class="table-container">
+          <Table>
+            <TableHeader>
+              <TableRow>
+                <TableHead> Name </TableHead>
+                <TableHead> Status </TableHead>
+                <TableHead> Due Date </TableHead>
+              </TableRow>
+            </TableHeader>
+            <TableBody>
+              <TableRow v-for="task in project.tasks" :key="task.id">
+                <TableCell class="p-0">
+                  <RouterLink
+                    class="text-left block hover:bg-muted p-4"
+                    :to="{ name: '/tasks/[id]', params: { id: task.id } }"
+                  >
+                    {{ task.name }}
+                  </RouterLink>
+                </TableCell>
+                <TableCell>
+                  <AppInPlaceEditStatus readonly :modelValue="task.status" />
+                </TableCell>
+                <TableCell> {{ task.due_date }} </TableCell>
+              </TableRow>
+            </TableBody>
+          </Table>
+        </div>
       </div>
-    </div>
-    <div class="flex-1">
-      <h2>Documents</h2>
-      <div class="table-container">
-        <p class="text-muted-foreground text-sm font-semibold px-4 py-3">
-          This project doesn't have documents yet...
-        </p>
-        <!-- <Table>
+      <div class="flex-1">
+        <h2>Documents</h2>
+        <div class="table-container">
+          <p class="text-muted-foreground text-sm font-semibold px-4 py-3">
+            This project doesn't have documents yet...
+          </p>
+          <!-- <Table>
           <TableHeader>
             <TableRow>
               <TableHead> Name </TableHead>
@@ -99,12 +122,13 @@ await getProjects()
             </TableRow>
           </TableBody>
         </Table> -->
+        </div>
       </div>
-    </div>
-  </section>
+    </section>
+  </div>
 </template>
 
-<style>
+<style scoped>
 th {
   @apply w-[100px];
 }
