@@ -1,23 +1,21 @@
 <script setup lang="ts">
-import { tasksWithProjectsQuery } from '@/utils/supaQueries'
 import { columns } from '@/utils/tableColumns/tasksColumns'
-import type { TasksWithProjects } from '@/utils/supaQueries'
-import { useErrorStore } from '@/stores/error'
 
 usePageStore().pageData.title = 'My Tasks'
 
-const tasks = ref<TasksWithProjects | null>(null)
-const getTasks = async () => {
-  const { data, error, status } = await tasksWithProjectsQuery
-
-  if (error) useErrorStore().setError({ error, customCode: status })
-
-  tasks.value = data
-}
+const tasksLoader = useTasksStore()
+const { tasks } = storeToRefs(tasksLoader)
+const { getTasks } = tasksLoader
 
 await getTasks()
+
+const { getGroupedCollabs, groupedCollabs } = useCollabs()
+
+getGroupedCollabs(tasks.value ?? [])
+
+const columnsWithCollabs = columns(groupedCollabs)
 </script>
 
 <template>
-  <DataTable v-if="tasks" :columns="columns" :data="tasks" />
+  <DataTable v-if="tasks" :columns="columnsWithCollabs" :data="tasks" />
 </template>
