@@ -1,8 +1,13 @@
 import type { ColumnDef } from '@tanstack/vue-table'
 import type { TasksWithProjects } from '../supaQueries'
 import { RouterLink } from 'vue-router'
+import type { GroupedCollabs } from '@/types/GroupedCollabs'
+import Avatar from '@/components/ui/avatar/Avatar.vue'
+import AvatarImage from '@/components/ui/avatar/AvatarImage.vue'
+import AvatarFallback from '@/components/ui/avatar/AvatarFallback.vue'
+import AppInPlaceEditStatus from '@/components/AppInPlaceEdit/AppInPlaceEditStatus.vue'
 
-export const columns: ColumnDef<TasksWithProjects[0]>[] = [
+export const columns = (collabs: Ref<GroupedCollabs>): ColumnDef<TasksWithProjects[0]>[] => [
   {
     accessorKey: 'name',
     header: () => h('div', { class: 'text-left' }, 'Name'),
@@ -21,7 +26,14 @@ export const columns: ColumnDef<TasksWithProjects[0]>[] = [
     accessorKey: 'status',
     header: () => h('div', { class: 'text-left' }, 'Status'),
     cell: ({ row }) => {
-      return h('div', { class: 'text-left font-medium' }, row.getValue('status'))
+      return h(
+        'div',
+        { class: 'text-left font-medium' },
+        h(AppInPlaceEditStatus, {
+          modelValue: row.original.status,
+          readonly: true
+        })
+      )
     }
   },
   {
@@ -53,8 +65,18 @@ export const columns: ColumnDef<TasksWithProjects[0]>[] = [
     cell: ({ row }) => {
       return h(
         'div',
-        { class: 'text-left font-medium' },
-        JSON.stringify(row.getValue('collaborators'))
+        { class: 'text-left font-medium h-20 flex items-center' },
+        collabs.value[row.original.id]
+          ? collabs.value[row.original.id].map((collab) => {
+              return h(RouterLink, { to: `/users/${collab.username}` }, () => {
+                return h(Avatar, { class: 'hover:scale-110 transition-transform' }, () =>
+                  h(AvatarImage, { src: collab.avatar_url || '' })
+                )
+              })
+            })
+          : row.original.collaborators.map(() => {
+              return h(Avatar, { class: 'animate-pulse' }, () => h(AvatarFallback))
+            })
       )
     }
   }
